+++ packages/cloud-cognitive/src/components/Tearsheet/Tearsheet.stories.js
@@ -103,7 +103,7 @@ const closeIconDescription = 'Close the tearsheet';
 
 const description =
   'This is a description for the tearsheet, providing an opportunity to \
-  describe the flow.';
+  describe the flow over a couple of lines in the header of the tearsheet.';
 
 const influencer = (
   <div className="tearsheet-stories__dummy-content-block">Influencer</div>
@@ -116,11 +116,22 @@ const mainContent = (
     <Form>
       <p>Main content</p>
       <FormGroup legendText="">
-        <TextInput id="tss-ft1" labelText="Enter an important value here" />
+        <TextInput
+          id="tss-ft1"
+          labelText="Enter an important value here"
+          style={
+            // stylelint-disable-next-line carbon/layout-token-use
+            { marginBottom: '1em' }
+          }
+        />
         <TextInput
           id="tss-ft2"
           light
           labelText="Here is a light entry field:"
+          style={
+            // stylelint-disable-next-line carbon/layout-token-use
+            { marginBottom: '1em' }
+          }
         />
       </FormGroup>
     </Form>