+++ packages/cloud-cognitive/src/components/Tearsheet/TearsheetShell.js
@@ -196,7 +196,7 @@ export const TearsheetShell = React.forwardRef(
               })}
               iconDescription={closeIconDescription}>
               <Wrap className={`${bc}__header-content`}>
-                <Wrap>
+                <Wrap className={`${bc}__header-fields`}>
                   {/* we create the label and title here instead of passing them
                       as modal header props so we can wrap them in layout divs */}
                   <Wrap element="h2" className={`${bcModalHeader}__label`}>